Move mock and support functions to separate file for future sharing
